+++ model/src/main/java/com/aplana/sbrf/taxaccounting/model/raschsv/RaschsvVyplSvDopMt.java
@@ -2,6 +2,8 @@ package com.aplana.sbrf.taxaccounting.model.raschsv;
 
 import com.aplana.sbrf.taxaccounting.model.IdentityObject;
 
+import java.math.BigDecimal;
+
 /**
  * Сведения о сумме выплат и иных вознаграждений, исчисленных в пользу физического лица, на которые исчислены страховые взносы по дополнительному тарифу, по месяцу и коду тарифа (ВыплСВДопМТ)
  */
@@ -10,8 +12,8 @@ public class RaschsvVyplSvDopMt extends IdentityObject<Long> {
     private Long raschsvVyplSvDopId;
     private String mesyac;
     private String tarif;
-    private Double vyplSv;
-    private Double nachislSv;
+    private BigDecimal vyplSv;
+    private BigDecimal nachislSv;
 
     public static final String SEQ = "seq_raschsv_vypl_sv_dop_mt";
     public static final String TABLE_NAME = "raschsv_vypl_sv_dop_mt";
@@ -35,9 +37,9 @@ public class RaschsvVyplSvDopMt extends IdentityObject<Long> {
     public String getTarif() { return tarif; }
     public void setTarif(String tarif) { this.tarif = tarif; }
 
-    public Double getVyplSv() { return vyplSv; }
-    public void setVyplSv(Double vyplSv) { this.vyplSv = vyplSv; }
+    public BigDecimal getVyplSv() { return vyplSv; }
+    public void setVyplSv(BigDecimal vyplSv) { this.vyplSv = vyplSv; }
 
-    public Double getNachislSv() { return nachislSv; }
-    public void setNachislSv(Double nachislSv) { this.nachislSv = nachislSv; }
+    public BigDecimal getNachislSv() { return nachislSv; }
+    public void setNachislSv(BigDecimal nachislSv) { this.nachislSv = nachislSv; }
 }
